Fixed casting warnings.

+++ include/gameboard.h
@@ -153,8 +153,8 @@ namespace Gameboard
             IntVector2 dimensions = {};
             T_Entity square = m_grid[0][0];
 
-            dimensions.x = m_grid[0].size() * square.GetWidth() + (m_grid[0].size() - 1) * square.GetMarginWidth();
-            dimensions.y = m_grid.size() * square.GetHeight() + (m_grid.size() - 1) * square.GetMarginHeight();
+            dimensions.x = (int)m_grid[0].size() * square.GetWidth() + ((int)m_grid[0].size() - 1) * square.GetMarginWidth();
+            dimensions.y = (int)m_grid.size() * square.GetHeight() + ((int)m_grid.size() - 1) * square.GetMarginHeight();
 
             return dimensions;
         }
